Modularize the existing code into the `iam` package
